requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
isodate==0.7.2
lxml==5.4.0
oauthlib==3.2.2
platformdirs==4.3.7
pycparser==2.22
PyJWT==2.9.0
pyspnego==0.11.2
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
zeep==4.3.1
